Split hybrid_extractor.py: move regex patterns to patterns.py
hybrid_extractor.py was 1651 lines — the biggest file in the codebase
by a wide margin. Splitting the HybridExtractor class itself would be
risky (tightly coupled, F1-quality-gated, 93% tested — a mechanical
split could quietly regress extraction quality in ways unit tests
wouldn't catch), so this moves only the genuinely safe part: the ~780
lines of module-level regex constants and small pure text-analysis
helper functions (_clip, _is_negated_context, _tech_mention_is_a_decision,
etc.) that carry zero behavioral coupling to the class — they're data
and stateless functions, not pipeline logic.

Pure code motion done as a precise mechanical text split (not manual
retyping — real risk of transcription errors across 780 lines of dense
regex), followed by an explicit import block back into
hybrid_extractor.py so every name that used to be defined there
directly is still an attribute of that module (external code doing
`from tokenmizer.graph_memory.hybrid_extractor import _clip`, as
test_extraction_quality.py does, is unaffected). ruff caught 15 names
in my first draft of that import list that turned out to be used only
internally within patterns.py (e.g. _CLAUSE_SPAN, consumed by
_DECISION's own re.compile() call, never referenced by the class) —
removed rather than force-exported, since nothing outside patterns.py
needs them.

Verified zero behavior change, not just "looks fine": ran
benchmarks.eval before and after — every number identical (macro F1
94%, per-category P/R/F1, label-quality stats, all to the digit) — plus
both benchmark scripts and the full test suite, all unchanged.

Added tests/unit/test_patterns_split.py: guards the re-export contract
(the one name external code imports directly still resolves) and that
hybrid_extractor/patterns don't hold two independently-recompiled
copies of the same pattern that could silently drift apart later.

637 tests (+2), ruff clean.
